refactor(iwork): split the Pages backend into a model and two readers
The backend module had grown to 2644 lines and past the 1500-line limit the
repository enforces, with three unrelated things in it: what a Pages
document holds, how each of the two containers spells that, and how it
becomes a DoclingDocument.

They are separate modules now, which is also where the review of #3934
pointed:

- content: the model both generations read into, and the helpers that work
  on it — runs, paragraphs, lists, pictures, comments.
- pages_iwa: the Pages 5+ object graph, and only the message and field
  numbers as format knowledge.
- pages_xml: the iWork '09 sf: vocabulary.
- pages_backend: the container dispatch and the document assembly.

Each reader now exposes one read_content(), so the backend decides which
generation wrote the file and nothing else. No behaviour changes: every
Pages document available produces the same output, item for item, as
before the split.
